Dengan QR Generate

+++ json/admin_pegawai_aksi.php
@@ -14,6 +14,27 @@ if($_GET['page']=='simpan')
 	$tgl_lhr=$_POST['tgl_lahir'];
 	$alamat=$_POST['alamat'];
 	$no_telp=$_POST['no_telp'];
+/*
+	//CEK KEPSEK
+	$sql_cek_1 = "SELECT id_jbt FROM pegawai WHERE id_jbt='1'";
+	if(!$result = $db->query($sql_cek_1)){
+			die('There was an error running the query [' . $db->error . ']');
+			}
+	$kepsek_cek=$result->num_rows;
+	$a = '';
+	if ($kepsek_cek)
+	{$a = 1;}
+
+	//CEK WAKASEK
+	$sql_cek_2 = "SELECT id_jbt FROM pegawai WHERE id_jbt='2'";
+	if(!$result = $db->query($sql_cek_2)){
+			die('There was an error running the query [' . $db->error . ']');
+			}
+	$wakasek_cek=$result->num_rows;
+	$b = '';
+	if ($wakasek_cek)
+	{$b = 2;}
+*/
 
 	if (!empty($_FILES['foto']['name']))
 	{
@@ -54,7 +75,7 @@ if($_GET['page']=='simpan')
 				}					
 				else if ($ukuran_file>(1024*1024))
 				{
-					echo "Maksimal File Upload 1Mb !!! <br>";
+					echo json_encode(array('msg'=>'Maksimal File Upload 1Mb !!!'));
 				}
 				else
 				{
@@ -139,7 +160,7 @@ if($_GET['page']=='simpan')
 				
 			else if ($ukuran_file>(1024*1024))
 			{
-				echo "Maksimal File Upload 1Mb !!! <br>";
+				echo json_encode(array('msg'=>'Maksimal File Upload 1Mb !!!'));
 			}
 			else
 			{
@@ -225,36 +246,44 @@ if($_GET['page']=='edit')
 	$alamat=$_POST['alamat'];
 	$no_telp=$_POST['no_telp'];
 
-	if (!empty($_FILES['foto']['name']))
-	{
-		//CEK FILE JIKA ADA
-		$sql_cek = "SELECT foto 
-					FROM pegawai
-					WHERE id_peg='$id_peg'";
-		if(!$result = $db->query($sql_cek)){
+	$sql_cek = "SELECT nip, id_peg 
+				FROM pegawai
+				WHERE nip='$nip' && nip!='' && id_peg!='$id_peg'";
+	if(!$result = $db->query($sql_cek)){
 			die('There was an error running the query [' . $db->error . ']');
 			}
-		$id_peg_cek=$result->fetch_assoc();
-		$foto_cek = $id_peg_cek['foto'];
-		if (!empty($foto_cek) && file_exists("../foto/$foto_cek"))
+	$nip_cek=$result->num_rows;
+
+	if (!$nip_cek)
+	{
+		if (!empty($_FILES['foto']['name']))
 		{
-			//HAPUS FILE
-			unlink("../foto/$foto_cek");
-		}
+			//CEK FILE JIKA ADA
+			$sql_cek = "SELECT foto 
+						FROM pegawai
+						WHERE id_peg='$id_peg'";
+			if(!$result = $db->query($sql_cek)){
+				die('There was an error running the query [' . $db->error . ']');
+				}
+			$id_peg_cek=$result->fetch_assoc();
+			$foto_cek = $id_peg_cek['foto'];
+			if (!empty($foto_cek) && file_exists("../foto/$foto_cek"))
+			{
+				//HAPUS FILE
+				unlink("../foto/$foto_cek");
+			}
 
-		$tipe_file   = $_FILES['foto']['type'];
-		$lokasi_file = $_FILES['foto']['tmp_name'];
-		$ukuran_file = $_FILES['foto']['size'];
-		$path 		 = $_FILES['foto']['name'];
-		$ext 		 = pathinfo($path, PATHINFO_EXTENSION);
-		$nama_nospace= str_replace(' ', '_', $nama);
-		$nama_file	 = $nama_nospace."-".$tgl_lhr.".".$ext;
-		$direktori 	 = "../foto/$nama_file";
-	}
+			$tipe_file   = $_FILES['foto']['type'];
+			$lokasi_file = $_FILES['foto']['tmp_name'];
+			$ukuran_file = $_FILES['foto']['size'];
+			$path 		 = $_FILES['foto']['name'];
+			$ext 		 = pathinfo($path, PATHINFO_EXTENSION);
+			$nama_nospace= str_replace(' ', '_', $nama);
+			$nama_file	 = $nama_nospace."-".$tgl_lhr.".".$ext;
+			$direktori 	 = "../foto/$nama_file";
 
 
-	if (!empty($_FILES['foto']['name']))
-		{
+			//EKSTENSI
 			$ekstensi_diperbolehkan = array( "image/png", "image/jpg", "image/gif", "image/jpeg" );
 			if (!in_array($tipe_file, $ekstensi_diperbolehkan))
 			{
@@ -313,7 +342,11 @@ if($_GET['page']=='edit')
 				echo json_encode(array('msg'=>'Ada Masalah Penambahan data Pegawai '.$db->error));
 			}
 		}
-
+	}
+	else
+	{
+		echo json_encode(array('success'=>false,'msg'=>'NIP '.$nip.' Sudah ada'));
+	}
 
 }
 
@@ -353,6 +386,97 @@ if($_GET['page']=='hapus')
 	else
 	{
 		echo json_encode(array('msg'=>'Ada Masalah penghapusan pegawai '.$db->error));
+	}
+}
+
+if($_GET['page']=='qr_gen')
+{
+	include ('../admin/qr_lib_gen/qrlib.php');
+
+	$tempdir = "../admin/temp_qr/"; 
+	if (!file_exists($tempdir))
+	    mkdir($tempdir);
+
+	$id_peg=$_POST['id_peg'];
+	$nama=$_POST['nama'];
+	$qr_level=$_POST['qr_level'];
+	$qr_code1=$_POST['qr_code1'];
+	$qr_code2=$_POST['qr_code2'];
+	$isi_teks= $qr_code1.$qr_code2;
+	$namafile = $isi_teks.".png";
+	$quality = 'H'; //ada 4 pilihan, L (Low), M(Medium), Q(Good), H(High)
+	$ukuran = 5; //batasan 1 paling kecil, 10 paling besar
+	$padding = 1;
+
+	$sql_cek = "SELECT qr_code 
+				FROM barcode
+				WHERE qr_code='$isi_teks' && id_peg!='$id_peg'";
+	if(!$result = $db->query($sql_cek)){
+			die('There was an error running the query [' . $db->error . ']');
+			}
+	$qr_code_cek=$result->num_rows;
+
+	if (!$qr_code_cek)
+	{
+		if ($qr_level=='2')
+		{
+			$sql_cek = "SELECT b.qr_level, p.nama 
+						FROM barcode b
+						LEFT JOIN pegawai p ON p.id_peg=b.id_peg
+						WHERE qr_level='2'";
+
+			if(!$result = $db->query($sql_cek)){
+					die('There was an error running the query [' . $db->error . ']');
+					}
+
+			$level_cek=$result->fetch_assoc();
+			$nama_level_peg_cek = $level_cek['nama'];
+			if ($nama_level_peg_cek)
+			{
+				echo json_encode(array('success'=>false,'msg'=>'Level SPESIAL sudah ada. Yaitu pada: '.$nama_level_peg_cek));
+			}
+			else
+			{
+				$sql = $db->query("UPDATE barcode SET
+									qr_code		= '$isi_teks',
+									qr_level	= '$qr_level'
+									WHERE id_peg='$id_peg'");	
+
+						if($sql)
+						{
+							//GENERATE QR CODE
+							QRCode::png($isi_teks,$tempdir.$namafile,$quality,$ukuran,$padding);
+							echo json_encode(array('success'=>true,'namafile'=>$namafile,'nama'=>$nama));		
+						}
+						else
+						{
+							echo json_encode(array('msg'=>'Ada Masalah Pembaruan QR Code '.$db->error));
+						}
+			}
+		}
+		else
+		{
+			$sql = $db->query("UPDATE barcode SET
+									qr_code		= '$isi_teks',
+									qr_level	= '$qr_level'
+									WHERE id_peg='$id_peg'");	
+
+						if($sql)
+						{
+							//GENERATE QR CODE
+							QRCode::png($isi_teks,$tempdir.$namafile,$quality,$ukuran,$padding);
+
+							echo json_encode(array('success'=>true,'namafile'=>$namafile,'nama'=>$nama));		
+						}
+						else
+						{
+							echo json_encode(array('msg'=>'Ada Masalah Pembaruan QR Code '.$db->error));
+						}
+		}
+	}
+	else
+	{
+		echo json_encode(array('success'=>false,'msg'=>'QR Code '.$qr_code2.' sudah ada. Silakan ganti dengan yang lain'));
 	}
 }
